requirements.txt
amqp==5.3.1
anyio==3.7.1
asgiref==3.11.0
billiard==4.2.4
boto3==1.42.5
botocore==1.42.5
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.11
exceptiongroup==1.3.1
fastapi==0.75.2
greenlet==3.3.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
kombu==5.6.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==1.10.24
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzbar==0.1.9
qrcode==8.2
redis==4.5.5
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlparse==0.5.4
starlette==0.17.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.1
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
gunicorn
mysqlclient